requirements.txt
pytest-bdd
pytest
pytest-playwright
pytest-html
pytest-xdist
playwright
allure-pytest
requests
Pygments
setuptools
wheel